[1822] Migrate 1822 to new structure
